Reto1y2/Server/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic_core==2.16.3
httpx==0.19.0
